# manager-external-client/api/openapi.yaml
openapi: 3.0.0
info:
  description: Public API for the Preservation Manager.
  title: Preservation Manager Public API
  version: 1.0.0
servers:
- description: Local server
  url: http://localhost:8484/api
security:
- ApiKeyAuth: []
...
paths:
...
  /system/storage:
    get:
      operationId: getSystemStorage
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStorageResponse'
          description: Retrieves system wide storage details, including a list of
            all organizations in the system and their storage details
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Retrieves system wide storage details, including a list of all organizations
        in the system and their storage details
      tags:
      - PreservationManager
      x-accepts: application/json
  /system/ingestEventCounts:
    get:
      operationId: getIngestEventCounts
      parameters:
      - description: Array of event types.
        example: RECEIVE_BAG
        explode: false
        in: query
        name: eventType
        required: false
        schema:
          items:
            $ref: '#/components/schemas/EventType'
          type: array
        style: form
      - description: Array of event outcomes.
        example: SUCCESS_WITH_WARNINGS
        explode: false
        in: query
        name: eventOutcome
        required: false
        schema:
          items:
            $ref: '#/components/schemas/EventOutcome'
          type: array
        style: form
      - description: The starting date for the event counts.
        example: 2017-07-21
        explode: true
        in: query
        name: startDate
        required: true
        schema:
          format: date
          type: string
        style: form
      - description: The ending date for the event counts.
        example: 2017-07-21
        explode: true
        in: query
        name: endDate
        required: true
        schema:
          format: date
          type: string
        style: form
      - description: The ID of the organization.
        example: organization-id
        explode: true
        in: query
        name: organizationName
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemEventCountsResponse'
          description: System wide event counts, grouped by event outcome.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Counts of all ingest events in the system or organization if specified.
      tags:
      - PreservationManager
      x-accepts: application/json
  /system/preservationEventCounts:
    get:
      operationId: getPreservationEventCounts
      parameters:
      - description: Array of event types.
        example: RECEIVE_BAG
        explode: false
        in: query
        name: eventType
        required: false
        schema:
          items:
            $ref: '#/components/schemas/EventType'
          type: array
        style: form
      - description: Array of event outcomes.
        example: SUCCESS_WITH_WARNINGS
        explode: false
        in: query
        name: eventOutcome
        required: false
        schema:
          items:
            $ref: '#/components/schemas/EventOutcome'
          type: array
        style: form
      - description: The starting date for the event counts.
        example: 2017-07-21
        explode: true
        in: query
        name: startDate
        required: true
        schema:
          format: date
          type: string
        style: form
      - description: The ending date for the event counts.
        example: 2017-07-21
        explode: true
        in: query
        name: endDate
        required: true
        schema:
          format: date
          type: string
        style: form
      - description: The ID of the organization.
        example: organization-id
        explode: true
        in: query
        name: organizationName
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemEventCountsResponse'
          description: System wide event counts, grouped by event outcome.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Counts of all preservation events in the system or organization if
        specified.
      tags:
      - PreservationManager
      x-accepts: application/json
components:
  schemas:
...
    SystemStorageOrgDetails:
      example:
        objectCount: 10293
        orgName: organization-id
        displayName: Organization Name
        storageMegabytes: 10293
      properties:
        orgName:
          description: The ID of the organization
          example: organization-id
          type: string
        displayName:
          description: The display name of the organization
          example: Organization Name
          type: string
        objectCount:
          description: The total number of objects in the org
          example: 10293
          format: int64
          type: integer
        storageMegabytes:
          description: The approximate amount of space occupied by all object content
            files in the org in megabytes
          example: 10293
          format: int64
          type: integer
    SystemStorageResponse:
      example:
        totalObjectCount: 10293
        organizations:
        - objectCount: 10293
          orgName: organization-id
          displayName: Organization Name
          storageMegabytes: 10293
        - objectCount: 10293
          orgName: organization-id
          displayName: Organization Name
          storageMegabytes: 10293
        totalStorageMegabytes: 10293
      properties:
        totalObjectCount:
          description: The total number of objects in the system
          example: 10293
          format: int64
          type: integer
        totalStorageMegabytes:
          description: The approximate amount of space occupied by all object content
            files in the system in megabytes
          example: 10293
          format: int64
          type: integer
        organizations:
          description: Organization level storage details
          items:
            $ref: '#/components/schemas/SystemStorageOrgDetails'
          type: array
    EventOutcomeCount:
      example:
        approved: 10293
        rejected: 10293
        success: 10293
        failure: 10293
        successWithWarnings: 10293
        notExecuted: 10293
      properties:
        success:
          description: Count of events with outcome 'SUCCESS'.
          example: 10293
          format: int64
          type: integer
        failure:
          description: Count of events with outcome 'FAILURE'.
          example: 10293
          format: int64
          type: integer
        successWithWarnings:
          description: Count of events with outcome 'SUCCESS_WITH_WARNINGS'.
          example: 10293
          format: int64
          type: integer
        approved:
          description: Count of events with outcome 'APPROVED'.
          example: 10293
          format: int64
          type: integer
        rejected:
          description: Count of events with outcome 'REJECTED'.
          example: 10293
          format: int64
          type: integer
        notExecuted:
          description: Count of events with outcome 'NOT_EXECUTED'.
          example: 10293
          format: int64
          type: integer
    EventCounts:
      example:
        eventCountsByOutcome:
          approved: 10293
          rejected: 10293
          success: 10293
          failure: 10293
          successWithWarnings: 10293
          notExecuted: 10293
      properties:
        eventType:
          $ref: '#/components/schemas/EventType'
        eventCountsByOutcome:
          $ref: '#/components/schemas/EventOutcomeCount'
    SystemEventCountsResponse:
      example:
        eventCounts:
        - eventCountsByOutcome:
            approved: 10293
            rejected: 10293
            success: 10293
            failure: 10293
            successWithWarnings: 10293
            notExecuted: 10293
        - eventCountsByOutcome:
            approved: 10293
            rejected: 10293
            success: 10293
            failure: 10293
            successWithWarnings: 10293
            notExecuted: 10293
      properties:
        eventCounts:
          items:
            $ref: '#/components/schemas/EventCounts'
          type: array
...
    ErrorResponse:
      properties:
        timestamp:
          description: Timestamp when the error occurred
          format: date-time
          type: string
        requestId:
          description: UUID assigned to the request
          example: 6037de0a4b2e4611956da540938a00ff
          type: string
        status:
          description: HTTP status code
          example: 500
          format: int32
          type: integer
        error:
          description: Application classification of the error
          type: string
        message:
          description: Description of the error
          example: Invalid input
          type: string
        path:
          description: Request path
          example: /api/ingest/bag
          type: string
      required:
      - error
      - message
      - path
      - requestId
      - status
      - timestamp
    EventType:
      description: The type of event
      enum:
      - RECEIVE_BAG
      - VIRUS_SCAN_BAG
      - VALIDATE_BAG
      - IDENTIFY_OBJ
      - IDENTIFY_FILE_FORMAT
      - REVIEW_OBJ
      - REVIEW_BATCH
      - WRITE_OBJ_LOCAL
      - CREATE_OBJ
      - UPDATE_OBJ
      - UPDATE_OBJ_METADATA
      - REPLICATE_OBJ_VERSION
      - COMPLETE_OBJ_INGEST
      - COMPLETE_BATCH_INGEST
      - DELETE_BAG
      - VALIDATE_OBJ_LOCAL
      - VALIDATE_OBJ_VERSION_LOCAL
      - VALIDATE_OBJ_VERSION_REMOTE
      - RESTORE_OBJ_VERSION
      - PREPARE_DIP
      - DELETE_OBJ
      type: string
    EventOutcome:
      description: The outcome of an event
      enum:
      - SUCCESS
      - FAILURE
      - SUCCESS_WITH_WARNINGS
      - APPROVED
      - REJECTED
      - NOT_EXECUTED
      type: string
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: X-API-KEY
      type: apiKey